+++ src/core/RealtimeAnalyser.cpp
@@ -108,7 +108,7 @@ void RealtimeAnalyser::writeInput(ContextRenderLock &r, AudioBus* bus, size_t fr
     memcpy(dest, source, sizeof(float) * framesToProcess);
 
     // Sum all channels in one if numberOfChannels > 1.
-    unsigned int numberOfChannels = bus->numberOfChannels();
+    size_t numberOfChannels = bus->numberOfChannels();
     if (numberOfChannels > 1) {
         for (size_t i = 1; i < numberOfChannels; i++) {
             source = bus->channel(i)->data();
